flask_server/requirements.txt
alembic==1.14.0
bcrypt==4.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
pandas==2.1.0
requests==2.31.0 